requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0

# AutoGen dependencies (optional)
pyautogen>=0.2.0

# Additional utilities
numpy>=1.24.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 